backend-django/requirements.txt
django==5.0.*
djangorestframework==3.15.*
django-cors-headers==4.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
PyJWT==2.8.*
bcrypt==4.1.*
cryptography==42.*
gunicorn==21.*
